backend/requirements.txt
alembic==1.18.4
bcrypt==4.3.0
email-validator==2.3.0
fastapi==0.135.3
httpx==0.28.1
passlib==1.7.4
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
PyJWT==2.12.1
pytest==9.0.2
pytest-asyncio==1.3.0
python-multipart==0.0.22
redis==7.4.0
SQLAlchemy==2.0.48
uvicorn==0.42.0
